split up Pair.scala files
