+++ grimlock-core/src/main/scala/commbank/grimlock/framework/Codec.scala
@@ -22,6 +22,7 @@ import java.util.Date
 
 import scala.math.BigDecimal
 import scala.util.{ Success, Try }
+import scala.reflect.runtime.universe.TypeTag
 
 import shapeless.{ :+:, CNil, Coproduct }
 import shapeless.ops.coproduct.Inject
@@ -580,6 +581,69 @@ case object LongCodec extends Codec[Long] {
 
   private case object LongAsFloat extends (Long => Float) {
     def apply(l: Long): Float = l.toFloat
+  }
+}
+
+/** Codec for dealing with a pair of values. Each value must have a corresponding `Codec`. */
+case class PairCodec[
+X : TypeTag,
+Y : TypeTag
+](
+   xCodec: Codec[X],
+   yCodec: Codec[Y],
+   open: Char = '(',
+   separator: Char = ',',
+   close: Char = ')'
+ ) extends Codec[(X, Y)] { self =>
+  val converters: Set[Codec.Convert[(X, Y)]] = Set.empty
+  val date: Option[((X, Y)) => Date] = None
+  val integral: Option[Integral[(X, Y)]] = None
+  val numeric: Option[Numeric[(X, Y)]] = None
+  def ordering: Ordering[(X, Y)] = new Ordering[(X, Y)] {
+    def compare(x: (X, Y), y: (X, Y)): Int = self.compare(x, y)
+  }
+
+  def box(value: (X, Y)): Value[(X, Y)] = PairValue(value, this)
+
+  def compare(
+               x: (X, Y),
+               y: (X, Y)
+             ): Int = if (xCodec.compare(x._1, y._1) == 0) yCodec.compare(x._2, y._2) else xCodec.compare(x._1, y._1)
+
+  def decode(
+              str: String
+            ): Option[(X, Y)] = {
+    val pattern = s"\\${open}(.+)\\${separator}(.+)\\${close}".r
+    val (x, y) = str match {
+      case pattern(left, right) => (xCodec.decode(left), yCodec.decode(right))
+      case _ => (None, None)
+    }
+    for (a <- x; b <- y) yield (a, b)
+  }
+
+  def encode(value: (X, Y)): String = s"${open}${xCodec.encode(value._1)}${separator}${yCodec.encode(value._2)}${close}"
+
+  def toShortString = s"pair${separator.toString}${xCodec.toShortString}${separator.toString}${yCodec.toShortString}${separator.toString}${open.toString}${separator.toString}${close.toString}"
+}
+
+object PairCodec {
+  /** Pattern for parsing `PairCodec` from string with xCodec and yCodec provided. */
+  val Pattern = raw"pair(.)(.+)\1(.+)\1(.)\1(.)".r
+
+  /**
+    * Parse a PairCodec[X, Y] from a string.
+    *
+    * @param str String from which to parse the codec.
+    * @param xCodec Left Codec[X] to attempt to parse PairCodec.
+    * @param yCodec Left Codec[Y] to attempt to parse PairCodec.
+    *
+    * @return A `Some[PairCodec[X, Y]]` in case of success, `None` otherwise.
+    */
+  def fromShortString[X : TypeTag, Y : TypeTag](str: String, xCodec: Codec[X], yCodec: Codec[Y]): Option[PairCodec[X, Y]] = str match {
+    case Pattern(sep, x, y, open, close) if xCodec.toShortString == x && yCodec.toShortString == y =>
+      // sep, open, close are regex-ed to be single length strings. So we can safely take their first element
+      Option(PairCodec(xCodec, yCodec, open.head, sep.head, close.head))
+    case _ => None
   }
 }
